+++ CHANGES.rst
@@ -1,6 +1,7 @@
 v1.0b3
 ======
 
+* ADDED: DocumentListener shortcuts in swingutils.events
 * CHANGED: @inlineCallbacks now always returns an AsyncToken for consistency
 * CHANGED: @inlineCallbacks now uses @swingRun instead of @swingCall to avoid
   certain undesirable side effects (the calling thread would get stuck waiting