+++ CHANGES.rst
@@ -2,6 +2,7 @@ v1.0b3
 ======
 
 * ADDED: DocumentListener shortcuts in swingutils.events
+* FIXED: EmptyNumberFormatter raised an AttributeException
 * CHANGED: @inlineCallbacks now always returns an AsyncToken for consistency
 * CHANGED: @inlineCallbacks now uses @swingRun instead of @swingCall to avoid
   certain undesirable side effects (the calling thread would get stuck waiting